refactor(drivers): Restructure drivers directory
BREAKING CHANGE: The `drivers` directory has been removed from the project structure. All driver modules have been moved up one level higher to promote better organization and accessibility. For example, the `github.com/bartventer/gorm-multitenancy/drivers/postgres/v7` directory is now located at `github.com/bartventer/gorm-multitenancy/postgres/v7`.
