+++ postgres/docs/0001_intro_and_usage.md
@@ -1,6 +1,6 @@
 # PostgreSQL driver
 
-[![Go Reference](https://pkg.go.dev/badge/github.com/bartventer/gorm-multitenancy.svg)](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7)
+[![Go Reference](https://pkg.go.dev/badge/github.com/bartventer/gorm-multitenancy.svg)](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7)
 
 The PostgreSQL driver provides multitenancy support for PostgreSQL databases using `GORM`.
 
@@ -62,12 +62,12 @@ func (Book) IsTenantTable() bool {
 
 #### After DB Initialization
 
-Call [`RegisterModels`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#RegisterModels) after initializing the database to register all models.
+Call [`RegisterModels`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#RegisterModels) after initializing the database to register all models.
 
 ```go
 import (
     "gorm.io/gorm"
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 db, err := gorm.Open(postgres.New(postgres.Config{
@@ -81,12 +81,12 @@ err := postgres.RegisterModels(db, &Tenant{}, &Book{})
 
 #### During DB Initialization
 
-Alternatively, you can pass the models as variadic arguments to [`postgres.New`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#New) when creating the dialect.
+Alternatively, you can pass the models as variadic arguments to [`postgres.New`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#New) when creating the dialect.
 
 ```go
 import (
     "gorm.io/gorm"
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 db, err := gorm.Open(postgres.New(postgres.Config{
@@ -97,12 +97,12 @@ if err != nil {
 }
 ```
 
-Or pass the models as variadic arguments to [`postgres.Open`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#Open) when creating the dialect.
+Or pass the models as variadic arguments to [`postgres.Open`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#Open) when creating the dialect.
 
 ```go
 import (
     "gorm.io/gorm"
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 db, err := gorm.Open(postgres.Open(dsn, &Tenant{}, &Book{}), &gorm.Config{})
@@ -117,11 +117,11 @@ After all models have been [registered](#model-registration), we can perform tab
 
 #### Public Tables
 
-Call [`MigratePublicSchema`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#MigratePublicSchema) to create the public schema and migrate all public models.
+Call [`MigratePublicSchema`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#MigratePublicSchema) to create the public schema and migrate all public models.
 
 ```go
 import (
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 err := postgres.MigratePublicSchema(db)
@@ -129,11 +129,11 @@ err := postgres.MigratePublicSchema(db)
 
 #### Tenant Tables
 
-Call [`CreateSchemaForTenant`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#CreateSchemaForTenant) to create the schema for a tenant and migrate all tenant-specific models.
+Call [`CreateSchemaForTenant`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#CreateSchemaForTenant) to create the schema for a tenant and migrate all tenant-specific models.
 
 ```go
 import (
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 err := postgres.CreateSchemaForTenant(db, tenantSchemaName)
@@ -141,11 +141,11 @@ err := postgres.CreateSchemaForTenant(db, tenantSchemaName)
 
 ### Dropping Tenant Schemas
 
-Call [`DropSchemaForTenant`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#DropSchemaForTenant) to drop the schema and cascade all schema tables.
+Call [`DropSchemaForTenant`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#DropSchemaForTenant) to drop the schema and cascade all schema tables.
 
 ```go
 import (
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
 )
 
 err := postgres.DropSchemaForTenant(db, tenantSchemaName)
@@ -155,13 +155,13 @@ err := postgres.DropSchemaForTenant(db, tenantSchemaName)
 
 Conforming to the above conventions, foreign key constraints between public and tenant-specific models can be created just as if you were using a shared database and schema.
 
-You can embed the [postgres.TenantModel](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7#TenantModel) struct in your tenant model to add the necessary fields for the tenant model.
+You can embed the [postgres.TenantModel](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7#TenantModel) struct in your tenant model to add the necessary fields for the tenant model.
 
 Then create a foreign key constraint between the public and tenant-specific models using the `SchemaName` field as the foreign key.
 
 ```go
 import (
-    "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7"
+    "github.com/bartventer/gorm-multitenancy/postgres/v7"
     "gorm.io/gorm"
 )
 
@@ -193,7 +193,7 @@ func (Book) TableName() string {
 
 #### `WithTenantSchema`
 
-Use the [`WithTenantSchema`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7/scopes#WithTenantSchema) scope function when you want to perform operations on a tenant specific table, which may include foreign key constraints to a public schema table(s).
+Use the [`WithTenantSchema`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7/scopes#WithTenantSchema) scope function when you want to perform operations on a tenant specific table, which may include foreign key constraints to a public schema table(s).
 
 ```go
 db.Scopes(WithTenantSchema(tenantID)).Find(&Book{})
@@ -201,11 +201,11 @@ db.Scopes(WithTenantSchema(tenantID)).Find(&Book{})
 
 #### `SetSearchPath`
 
-Use the [`SetSearchPath`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/drivers/postgres/v7/schema#SetSearchPath) function when the tenant schema table has foreign key constraints you want to access belonging to other tables in the same tenant schema (and or foreign key relations to public tables).
+Use the [`SetSearchPath`](https://pkg.go.dev/github.com/bartventer/gorm-multitenancy/postgres/v7/schema#SetSearchPath) function when the tenant schema table has foreign key constraints you want to access belonging to other tables in the same tenant schema (and or foreign key relations to public tables).
 
 ```go
 import (
-    pgschema "github.com/bartventer/gorm-multitenancy/drivers/postgres/v7/schema"
+    pgschema "github.com/bartventer/gorm-multitenancy/postgres/v7/schema"
     "gorm.io/gorm"
 )
 db, resetSearchPath := pgschema.SetSearchPath(db, tenantSchemaName)
