+++ postgres/migrator_test.go
@@ -6,7 +6,7 @@ import (
 	"sync"
 	"testing"
 
-	"github.com/bartventer/gorm-multitenancy/drivers/postgres/v7/internal/testutil"
+	"github.com/bartventer/gorm-multitenancy/postgres/v7/internal/testutil"
 	"gorm.io/driver/postgres"
 	"gorm.io/gorm"
 )
